Added location, pubs and restauraunts full stack to team details screen

+++ app/src/main/java/com/example/theawayguide/repository/TeamRepository.kt
@@ -9,9 +9,9 @@ interface TeamRepository {
 
     suspend fun getTeamDetails(url: String): Team
 
-    suspend fun getRestaurants(latitude: Int, longitude: Int, radius: Int): List<Attraction>?
+    suspend fun getRestaurants(latitude: Double, longitude: Double, radius: Int): List<Attraction>?
 
-    suspend fun getHotels(latitude: Int, longitude: Int, radius: Int): List<Attraction>?
+    suspend fun getHotels(latitude: Double, longitude: Double, radius: Int): List<Attraction>?
 
-    suspend fun getPubs(latitude: Int, longitude: Int, radius: Int): List<Attraction>?
+    suspend fun getPubs(latitude: Double, longitude: Double, radius: Int): List<Attraction>?
 }
